Place search tests with search subsystem
