+++ crates/truthdb-core/src/engine/test_support.rs
@@ -1,0 +1,25 @@
+use std::path::PathBuf;
+
+use crate::storage::StorageOptions;
+
+pub(in crate::engine) fn test_storage_options() -> StorageOptions {
+    StorageOptions {
+        size_gib: 1,
+        wal_ratio: 0.05,
+        metadata_ratio: 0.08,
+        snapshot_ratio: 0.02,
+        allocator_ratio: 0.02,
+        reserved_ratio: 0.17,
+        default_collation: None,
+    }
+}
+
+pub(in crate::engine) fn unique_temp_path(label: &str) -> PathBuf {
+    let mut path = std::env::temp_dir();
+    let nanos = std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .expect("system time before unix epoch")
+        .as_nanos();
+    path.push(format!("truthdb-{label}-{nanos}.db"));
+    path
+}
